miner
Would eat up loads of RAM using the "findOre" function. Just crazy amounts, so I put in a simple, click ore, hardcode.
The only problem here in the hardcode is there is no room for an evade function if attacked by a random event,
I don't know how to pur in a "logout" function if the character has died or has been taken to a random even,
I thought if it hasn't activated dropLogs() for 20 while (true) cycles IN A ROW then that would activate logOut (),
But as you can see from woodcutter, I have no idea how to stop a loop when something in the loop has repeated itself 5 times in a row,
All I end up doing is when that something in the loop has done its thing 5 times, whether it was in a row or not then it stops the loop,
for some reason in this miner it would not find color in screenshot about 100 times and then finally find it. The screenshot was made small to just fit the ores.

And so for this reason, just to find an ore it would loop 100 times (even when I put the pixel count up to 30,000 for the screenshot) and then if there really was no color there, it would not find color in screenshot about 100000000 times and just ram would go 1gb+ so fast and crash my machine before I could click STOP on VSC because I only have 3gb RAM and running RS and VSC takes about 1.7gb. 

I tried also to run the code made on VSC on notepadqq, because notpadqq takes up less RAM, but I don't know how to do that, I click run and it gives me a wierd menu with a bunch of %url% and %url% path?